move copy helix job-* folder code to benchmarks_ci.py
