+++ scripts/benchmarks_ci.py
@@ -36,6 +36,10 @@ from subprocess import Popen, CalledProcessError
 import dotnet
 import micro_benchmarks
 
+# for temporary diagnostic, should br removed before PR.
+import shutil
+import platform
+
 def init_tools(
         architecture: str,
         dotnet_versions: str,
@@ -171,6 +175,13 @@ def __process_arguments(args: list):
     add_arguments(parser)
     return parser.parse_args(args)
 
+# diagnostic function, remove before PR. 
+def copyjob(src, dst, symlinks=False, ignore=None):
+    for item in os.listdir(src):
+        s = os.path.join(src, item)
+        d = os.path.join(dst, item)
+        if os.path.isdir(s) and "Job-" in s:
+            shutil.copytree(s, d, symlinks, ignore)
 
 def __main(args: list) -> int:
     validate_supported_runtime()
@@ -265,6 +276,10 @@ def __main(args: list) -> int:
             import upload
             upload.upload(globpath, upload_container, UPLOAD_QUEUE, UPLOAD_TOKEN_VAR, UPLOAD_STORAGE_URI)
         # TODO: Archive artifacts.
+
+    # diagnostic function, remove before PR. 
+    if platform.system() == "Linux":
+        copyjob('%HELIX_CORRELATION_PAYLOAD\performance\artifacts\bin\MicroBenchmarks\Release\net6.0', '%HELIX_WORKITEM_UPLOAD_ROOT\BDNBackUp')
 
 
 if __name__ == "__main__":
